+++ specs/bitId/bitId.spec.js
@@ -6,8 +6,8 @@ describe('Bit-id', () => {
     it('should convert CSSComp to css-comp', () => {
       const bitName = 'CSSComp';
       expect(BitId.getValidBitId('global', bitName).toString()).to.equal('global/css-comp');
-    });
-    it('should convert CSSComp to en-us', () => {
+    }).timeout(5000);
+    it('should convert EN-US to en-us', () => {
       const bitName = 'EN-US';
       expect(BitId.getValidBitId('global', bitName).toString()).to.equal('global/en-us');
     });