+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - improve bit add to convert files to valid bit names 
 - improve checkVersionCompatibility between server and client
+- show correct message / error when the tester has an exception during bit test
 
 ## [0.10.9-dev.1] - 2017-10-02